Removed a not useful exception, smaller fixes to router (honoring is Browseable too) and fix for NEXUS-1537
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@3274 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-proxy/src/main/java/org/sonatype/nexus/proxy/repository/AbstractProxyRepository.java
@@ -524,7 +524,10 @@ public abstract class AbstractProxyRepository
                             // cleanup if any remnant is here
                             try
                             {
-                                deleteItem( uid, context );
+                                if ( localItem == null )
+                                {
+                                    deleteItem( uid, context );
+                                }
                             }
                             catch ( ItemNotFoundException ex1 )
                             {
